update inv cal
