+++ override_price_calculation/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 # See LICENSE file for full copyright and licensing details.
-from . import override_price_calculation
+from . import override_price_calculation_sales, override_price_calculation_inv
